Ai - basis implementatie ai
