+++ .env
@@ -2,3 +2,4 @@ MYSQL_HOST=mysql
 MYSQL_USERNAME=root
 MYSQL_PASSWORD=password
 MYSQL_DATABASE=hive
+AI_API_HOST=ai:5000
