--- app/.env
@@ -1,4 +1,0 @@
-MYSQL_HOST=mysql
-MYSQL_USERNAME=root
-MYSQL_PASSWORD=password
-MYSQL_DATABASE=hive
