+++ app/src/Player.php
@@ -2,10 +2,14 @@
 
 namespace app;
 
+use app\ai\Ai;
+
 class Player
 {
     private array $hand;
     private int $playerNumber;
+
+    private Ai $ai;
 
     public function __construct(
         $playerNumber,
@@ -37,6 +41,21 @@ class Player
         } else {
             $this->hand[$piece]--;
         }
+    }
+
+    public function getAi(): Ai
+    {
+        return $this->ai;
+    }
+
+    public function setAi(Ai $ai): void
+    {
+        $this->ai = $ai;
+    }
+
+    public function isAi(): bool
+    {
+        return $this->getAi() !== null;
     }
 
 }
