+++ app/src/formPosts/play.php
@@ -2,6 +2,7 @@
 
 require_once '../../vendor/autoload.php';
 
+use app\ai\CurlRequest;
 use app\Game;
 use app\Database;
 use app\Moves;
@@ -17,5 +18,12 @@ $piece = $_POST['piece'];
 $toPosition = $_POST['toPosition'];
 
 Moves::playPiece($piece, $toPosition, $game, $database);
+if ($game->getCurrentPlayer()->isAi()) {
+    $url = $_ENV["AI_API_HOST"];
+    $curlRequest = new CurlRequest($url);
+    $ai = $game->getCurrentPlayer()->getAi();
+    $ai->aiPlaysTurn($game, $database, $curlRequest);
+}
 
-header('Location: /../../index.php');
+
+header('Location: /../../main.php');
